ANIMATION FUNCTIONALITY DONE!
FINALLY! The animation functionality is done. Had some trouble with JQuery because I'm completely new to it. As is nearly all code error, I had some syntax stuff wrong and weird things not playing nicely with one another. Now it's fixed: animation.js contains functions which switch the classes on text and motion graphic to make them appear and be equal size. Wahoo!